add python realtime speech transcription + select ai nl2sql
