+++ python-realtimespeech-selectai/requirements.txt
@@ -3,3 +3,4 @@ https://artifactory.oci.oraclecorp.com/global-dev-pypi/oci-2.129.1+preview.1.180
 PyAudio>=0.2.14
 websockets==11.0.3; python_version >= "3.7"
 websockets==9.1; python_version == "3.6"
+oracledb
